impl for #566
