+++ src/main/java/cyclops/Streams.java
@@ -2,18 +2,17 @@ package cyclops;
 
 import com.aol.cyclops2.hkt.Higher;
 import com.aol.cyclops2.internal.stream.spliterators.*;
-import cyclops.box.LazyImmutable;
+import cyclops.collections.box.LazyImmutable;
 import cyclops.collections.immutable.PVectorX;
 import cyclops.function.*;
 import cyclops.monads.AnyM;
 import cyclops.stream.ReactiveSeq;
 import cyclops.stream.Streamable;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import com.aol.cyclops2.data.collections.extensions.CollectionX;
 import cyclops.collections.ListX;
 import com.aol.cyclops2.internal.stream.*;
 import com.aol.cyclops2.internal.stream.operators.*;
-import com.aol.cyclops2.internal.stream.spliterators.push.CapturingOperator;
 import cyclops.monads.Witness;
 import cyclops.monads.Witness.stream;
 import com.aol.cyclops2.types.stream.HeadAndTail;
