+++ src/main/java/cyclops/collections/box/LazyImmutable.java
@@ -1,4 +1,4 @@
-package cyclops.box;
+package cyclops.collections.box;
 
 import com.aol.cyclops2.types.Transformable;
 import cyclops.stream.ReactiveSeq;
