+++ src/main/java/cyclops/collections/box/MutableChar.java
@@ -1,4 +1,4 @@
-package cyclops.box;
+package cyclops.collections.box;
 
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -17,7 +17,7 @@ import lombok.ToString;
  * variables can't be changed.
  * e.g.
  *<pre>{@code 
- * byte var = true;
+ * char var = true;
  * Runnable r = () -> var =false;
  * }</pre>
  * 
@@ -26,7 +26,7 @@ import lombok.ToString;
  * 
  * e.g.
  * <pre>{@code
- * MutableByte var =  MutableByte.of(true);
+ * MutableChar var =  MutableChar.of(true);
  * Runnable r = () -> var.set(false);
  * }</pre>
  * 
@@ -38,60 +38,60 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 @EqualsAndHashCode
-public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byte> {
+public class MutableChar implements To<MutableChar>,Supplier<Character>, Consumer<Character> {
 
-    private byte var;
+    private char var;
 
     /**
      * Create a Mutable variable, which can be mutated inside a Closure 
      * 
      * e.g.
      * <pre>{@code
-     *   MutableByte num = MutableByte.of(true);
+     *   MutableChar char = MutableChar.of('c');
      *   
-     *    num.mutate(n->!n))
+     *    char.mutate(n->'d'))
      *   
-     *   System.out.println(num.getAsByte());
-     *   //prints false
+     *   System.out.println(num.getAsChar());
+     *   //prints d
      * } </pre>
      * 
      * @param var Initial value of Mutable
      * @return New Mutable instance
      */
-    public static <T> MutableByte of(final byte var) {
-        return new MutableByte(
+    public static <T> MutableChar of(final char var) {
+        return new MutableChar(
                                var);
     }
 
     /** 
-     * Construct a MutableByte that gets and sets an external value using the provided Supplier and Consumer
+     * Construct a MutableChar that gets and sets an external value using the provided Supplier and Consumer
      * 
      * e.g.
      * <pre>
      * {@code 
-     *    MutableByte mutable = MutableByte.fromExternal(()->!this.value,val->!this.value);
+     *    MutableChar mutable = MutableChar.fromExternal(()->!this.value,val->!this.value);
      * }
      * </pre>
      * 
      * 
      * @param s Supplier of an external value
      * @param c Consumer that sets an external value
-     * @return MutableByte that gets / sets an external (mutable) value
+     * @return MutableChar that gets / sets an external (mutable) value
      */
-    public static MutableByte fromExternal(final Supplier<Byte> s, final Consumer<Byte> c) {
-        return new MutableByte() {
+    public static MutableChar fromExternal(final Supplier<Character> s, final Consumer<Character> c) {
+        return new MutableChar() {
             @Override
-            public byte getAsByte() {
+            public char getAsChar() {
                 return s.get();
             }
 
             @Override
-            public Byte get() {
-                return getAsByte();
+            public Character get() {
+                return getAsChar();
             }
 
             @Override
-            public MutableByte set(final byte value) {
+            public MutableChar set(final char value) {
                 c.accept(value);
                 return this;
             }
@@ -102,10 +102,10 @@ public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byt
      * Use the supplied function to perform a lazy map operation when get is called 
      * <pre>
      * {@code 
-     *  MutableByte mutable = MutableByte.fromExternal(()->!this.value,val->!this.value);
-     *  Mutable<Byte> withOverride = mutable.mapOutputToObj(b->{ 
+     *  MutableChar mutable = MutableChar.fromExternal(()->!this.value,val->!this.value);
+     *  Mutable<Character> withOverride = mutable.mapOutput(b->{ 
      *                                                        if(override)
-     *                                                             return 1b;
+     *                                                             return 'a';
      *                                                         return b;
      *                                                         });
      *          
@@ -116,8 +116,8 @@ public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byt
      * @param fn Map function to be applied to the result when get is called
      * @return Mutable that lazily applies the provided function when get is called to the return value
      */
-    public <R> Mutable<R> mapOutputToObj(final Function<Byte, R> fn) {
-        final MutableByte host = this;
+    public <R> Mutable<R> mapOutputToObj(final Function<Character, R> fn) {
+        final MutableChar host = this;
         return new Mutable<R>() {
             @Override
             public R get() {
@@ -131,10 +131,10 @@ public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byt
      * Use the supplied function to perform a lazy map operation when get is called 
      * <pre>
      * {@code 
-     *  MutableByte mutable = MutablByte.fromExternal(()->!this.value,val->!this.value);
-     *  Mutable<Byte> withOverride = mutable.mapInputToObj(b->{ 
+     *  MutableChar mutable = MutableChar.fromExternal(()->!this.value,val->!this.value);
+     *  Mutable<Character> withOverride = mutable.mapInput(b->{ 
      *                                                        if(override)
-     *                                                             return 1b;
+     *                                                             return 'v';
      *                                                         return b;
      *                                                         });
      *          
@@ -145,8 +145,8 @@ public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byt
      * @param fn Map function to be applied to the input when set is called
      * @return Mutable that lazily applies the provided function when set is called to the input value
      */
-    public <T1> Mutable<T1> mapInputToObj(final Function<T1, Byte> fn) {
-        final MutableByte host = this;
+    public <T1> Mutable<T1> mapInputToObj(final Function<T1, Character> fn) {
+        final MutableChar host = this;
         return new Mutable<T1>() {
             @Override
             public Mutable<T1> set(final T1 value) {
@@ -160,7 +160,7 @@ public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byt
     /**
      * @return Current value
      */
-    public byte getAsByte() {
+    public char getAsChar() {
         return var;
     }
 
@@ -168,7 +168,7 @@ public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byt
      * @param var New value
      * @return  this object with mutated value
      */
-    public MutableByte set(final byte var) {
+    public MutableChar set(final char var) {
         this.var = var;
         return this;
     }
@@ -177,24 +177,24 @@ public class MutableByte implements To<MutableByte>,Supplier<Byte>, Consumer<Byt
      * @param varFn New value
      * @return  this object with mutated value
      */
-    public MutableByte mutate(final ByteFunction varFn) {
-        return set(varFn.apply(getAsByte()));
+    public MutableChar mutate(final CharFunction varFn) {
+        return set(varFn.apply(get()));
 
     }
 
-    public static interface ByteFunction {
-        byte apply(byte var);
+    public static interface CharFunction {
+        char apply(char var);
     }
 
     @Override
-    public void accept(final Byte t) {
+    public void accept(final Character t) {
         set(t);
 
     }
 
     @Override
-    public Byte get() {
-        return getAsByte();
+    public Character get() {
+        return getAsChar();
     }
 
 }
