+++ src/main/java/cyclops/collections/box/MutableLong.java
@@ -1,4 +1,4 @@
-package cyclops.box;
+package cyclops.collections.box;
 
 import java.util.OptionalLong;
 import java.util.function.*;
