+++ src/main/java/cyclops/control/Eval.java
@@ -3,9 +3,8 @@ package cyclops.control;
 import com.aol.cyclops2.data.collections.extensions.CollectionX;
 import com.aol.cyclops2.hkt.Higher;
 import com.aol.cyclops2.types.*;
-import com.aol.cyclops2.types.stream.reactive.ValueSubscriber;
 import cyclops.async.Future;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.DequeX;
 import cyclops.collections.ListX;
 import cyclops.collections.immutable.PVectorX;
@@ -14,7 +13,6 @@ import cyclops.monads.AnyM;
 import cyclops.monads.Witness;
 import cyclops.monads.WitnessType;
 import cyclops.monads.transformers.EvalT;
-import cyclops.monads.transformers.MaybeT;
 import cyclops.stream.ReactiveSeq;
 import cyclops.stream.Spouts;
 import cyclops.typeclasses.Pure;
@@ -24,7 +22,6 @@ import cyclops.typeclasses.functor.Functor;
 import cyclops.typeclasses.instances.General;
 import cyclops.typeclasses.monad.*;
 import lombok.AllArgsConstructor;
-import lombok.experimental.Delegate;
 import lombok.experimental.UtilityClass;
 import org.jooq.lambda.tuple.Tuple2;
 import org.jooq.lambda.tuple.Tuple3;
