+++ src/main/java/cyclops/function/FluentFunctions.java
@@ -3,7 +3,7 @@ package cyclops.function;
 import cyclops.monads.AnyM;
 import cyclops.stream.ReactiveSeq;
 import cyclops.control.Try;
-import cyclops.box.MutableInt;
+import cyclops.collections.box.MutableInt;
 import com.aol.cyclops2.internal.invokedynamic.CheckedTriFunction;
 import cyclops.monads.WitnessType;
 import com.aol.cyclops2.util.ExceptionSoftener;
