+++ src/test/java/com/aol/cyclops2/closures/mutable/MutableBooleanTest.java
@@ -11,8 +11,8 @@ import java.util.stream.Stream;
 
 import org.junit.Test;
 
-import cyclops.box.Mutable;
-import cyclops.box.MutableBoolean;
+import cyclops.collections.box.Mutable;
+import cyclops.collections.box.MutableBoolean;
 
 public class MutableBooleanTest {
