+++ src/test/java/com/aol/cyclops2/closures/mutable/MutableCharTest.java
@@ -11,8 +11,8 @@ import java.util.stream.Stream;
 
 import org.junit.Test;
 
-import cyclops.box.Mutable;
-import cyclops.box.MutableChar;
+import cyclops.collections.box.Mutable;
+import cyclops.collections.box.MutableChar;
 public class MutableCharTest {
 
 	@Test
