+++ src/test/java/com/aol/cyclops2/closures/mutable/MutableFloatTest.java
@@ -11,8 +11,8 @@ import java.util.stream.Stream;
 
 import org.junit.Test;
 
-import cyclops.box.Mutable;
-import cyclops.box.MutableFloat;
+import cyclops.collections.box.Mutable;
+import cyclops.collections.box.MutableFloat;
 public class MutableFloatTest {
 
 	@Test
