+++ src/test/java/com/aol/cyclops2/closures/mutable/MutableIntTest.java
@@ -11,8 +11,8 @@ import java.util.stream.Stream;
 
 import org.junit.Test;
 
-import cyclops.box.Mutable;
-import cyclops.box.MutableInt;
+import cyclops.collections.box.Mutable;
+import cyclops.collections.box.MutableInt;
 public class MutableIntTest {
 
 	@Test
