+++ src/test/java/com/aol/cyclops2/control/anym/BaseAnyMValueTest.java
@@ -7,7 +7,7 @@ import cyclops.control.*;
 import cyclops.function.Monoid;
 import cyclops.Reducers;
 import cyclops.Semigroups;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.ListX;
 import com.aol.cyclops2.types.anyM.AnyMValue;
 import cyclops.monads.WitnessType;
