+++ src/test/java/com/aol/cyclops2/control/anym/MaybeAnyMValueTest.java
@@ -2,7 +2,7 @@ package com.aol.cyclops2.control.anym;
 
 import cyclops.monads.AnyM;
 import cyclops.control.Maybe;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.monads.Witness;
 import org.junit.Before;
 import org.junit.Test;
