+++ src/test/java/com/aol/cyclops2/control/transformers/CompletableFutureTTest.java
@@ -7,7 +7,7 @@ import cyclops.Reducers;
 import cyclops.Semigroups;
 import cyclops.Streams;
 import cyclops.async.Future;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.ListX;
 import cyclops.collections.immutable.PStackX;
 import cyclops.control.Maybe;
