+++ src/test/java/com/aol/cyclops2/control/transformers/MaybeTTest.java
@@ -5,8 +5,7 @@ import com.aol.cyclops2.types.mixins.Printable;
 import cyclops.Reducers;
 import cyclops.Semigroups;
 import cyclops.Streams;
-import cyclops.async.Future;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.ListX;
 import cyclops.collections.immutable.PStackX;
 import cyclops.control.Maybe;
@@ -16,7 +15,6 @@ import cyclops.control.Xor;
 import cyclops.function.Monoid;
 import cyclops.monads.AnyM;
 import cyclops.monads.Witness;
-import cyclops.monads.transformers.FutureT;
 import cyclops.monads.transformers.MaybeT;
 import cyclops.stream.ReactiveSeq;
 import org.junit.Before;
