+++ src/test/java/cyclops/control/either/CompletableEither3Test.java
@@ -1,12 +1,11 @@
 package cyclops.control.either;
 
 
-import cyclops.Monoids;
 import cyclops.Reducers;
 import cyclops.Semigroups;
 import cyclops.Streams;
 import cyclops.async.Future;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.ListX;
 import cyclops.collections.immutable.PStackX;
 import cyclops.control.*;
