+++ src/test/java/cyclops/control/either/CompletableEither5Test.java
@@ -5,11 +5,10 @@ import cyclops.Reducers;
 import cyclops.Semigroups;
 import cyclops.Streams;
 import cyclops.async.Future;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.ListX;
 import cyclops.collections.immutable.PStackX;
 import cyclops.control.*;
-import cyclops.control.either.Either5.CompletableEither5;
 import cyclops.control.either.Either5.CompletableEither5;
 import cyclops.function.Monoid;
 import org.jooq.lambda.Seq;
