+++ src/test/java/cyclops/control/either/CompletableEitherTest.java
@@ -5,14 +5,12 @@ import cyclops.Reducers;
 import cyclops.Semigroups;
 import cyclops.Streams;
 import cyclops.async.Future;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.ListX;
-import cyclops.collections.immutable.PSetX;
 import cyclops.collections.immutable.PStackX;
 import cyclops.control.*;
 import cyclops.control.either.Either.CompletableEither;
 import cyclops.function.Monoid;
-import cyclops.stream.Spouts;
 import org.jooq.lambda.Seq;
 import org.jooq.lambda.tuple.Tuple;
 import org.junit.Before;
