+++ src/test/java/cyclops/control/either/EitherTest.java
@@ -5,7 +5,7 @@ import cyclops.Reducers;
 import cyclops.Semigroups;
 import cyclops.Streams;
 import cyclops.async.Future;
-import cyclops.box.Mutable;
+import cyclops.collections.box.Mutable;
 import cyclops.collections.ListX;
 import cyclops.collections.immutable.PSetX;
 import cyclops.collections.immutable.PStackX;
